backend/requirements.txt
fastapi
uvicorn
SQLAlchemy
bcrypt
PyJWT
python-dotenv
pydantic
pydantic[email]
python-multipart